Move DIC_NR to __init__ and remove output files
